# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_ssr0016_ssr-blog_internal_apperror.ErrorResponse:
    properties:
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
      message:
        type: string
    type: object
...
  github_com_ssr0016_ssr-blog_internal_model.PostResponse:
    properties:
      content:
        type: string
      cover_image_url:
        type: string
      created_at:
        type: string
      excerpt:
        type: string
      id:
        type: integer
      published_at:
        type: string
      slug:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  github_com_ssr0016_ssr-blog_internal_model.PostSummary:
    properties:
      cover_image_url:
        type: string
      excerpt:
        type: string
      published_at:
        type: string
      slug:
        type: string
      title:
        type: string
    type: object
...
  github_com_ssr0016_ssr-blog_pkg_pagination.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_ssr0016_ssr-blog_pkg_pagination.Response-github_com_ssr0016_ssr-blog_internal_model_PostSummary:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_ssr0016_ssr-blog_internal_model.PostSummary'
        type: array
      meta:
        $ref: '#/definitions/github_com_ssr0016_ssr-blog_pkg_pagination.Meta'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: ssrBlog - Personal blog with session auth and RBAC
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: ssrBlog API
  version: "1.0"
paths:
...
  /posts:
    get:
      description: Public. Returns summaries of published posts, newest first. Full
        content is not included; fetch a single post by slug for that. Out-of-range
        page and limit values are clamped.
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ssr0016_ssr-blog_pkg_pagination.Response-github_com_ssr0016_ssr-blog_internal_model_PostSummary'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssr0016_ssr-blog_internal_apperror.ErrorResponse'
      summary: List published blog posts
      tags:
      - posts
  /posts/{slug}:
    get:
      description: Public. A draft, a deleted post and a slug that does not exist
        all return the same 404 response.
      parameters:
      - description: Post slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ssr0016_ssr-blog_internal_model.PostResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ssr0016_ssr-blog_internal_apperror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssr0016_ssr-blog_internal_apperror.ErrorResponse'
      summary: Get a published blog post by slug
      tags:
      - posts
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: ssrblog_session
    type: apiKey
swagger: "2.0"